client: Move shortcuts to App
Also use tinykeys so that we can reduce jQuery usage.
