+++ assets/package.json
@@ -24,6 +24,7 @@
     "react-dom": "^17.0.1",
     "reset-css": "^5.0.1",
     "spectrum-colorpicker": "^1.8.1",
+    "tinykeys": "^1.1.1",
     "unreset-css": "^1.0.1"
   },
   "devDependencies": {
